@@ -1,9 +1,9 @@
 package com.economato.inventory.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -20,15 +20,15 @@ public abstract class BaseControllerMockTest {
     @Autowired
     protected MockMvc mockMvc;
 
-    @MockBean
+    @MockitoBean
     protected StockLedgerService stockLedgerService;
 
-    @MockBean
+    @MockitoBean
     protected InventoryAuditService inventoryAuditService;
 
-    @MockBean
+    @MockitoBean
     protected OrderAuditService orderAuditService;
 
-    @MockBean
+    @MockitoBean
     protected RecipeAuditService recipeAuditService;
 }
