@@ -2,7 +2,8 @@ package com.economato.inventory.config;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.jsontype.impl.LaissezFaireSubTypeValidator;
+import com.fasterxml.jackson.databind.jsontype.BasicPolymorphicTypeValidator;
+import com.fasterxml.jackson.databind.jsontype.PolymorphicTypeValidator;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
@@ -45,12 +46,14 @@ public class RedisConfig {
          * - allergens: 24 horas (datos maestros)
          */
         @Bean
-        public CacheManager cacheManager(RedisConnectionFactory connectionFactory) {
-                // ObjectMapper personalizado para serialización
-                ObjectMapper objectMapper = new ObjectMapper();
-                objectMapper.registerModule(new JavaTimeModule());
+        public CacheManager cacheManager(RedisConnectionFactory connectionFactory, ObjectMapper objectMapper) {
+                // Configuración de validación de tipos polimórficos
+                PolymorphicTypeValidator ptv = BasicPolymorphicTypeValidator.builder()
+                                .allowIfBaseType(Object.class)
+                                .build();
+
                 objectMapper.activateDefaultTyping(
-                                LaissezFaireSubTypeValidator.instance,
+                                ptv,
                                 ObjectMapper.DefaultTyping.NON_FINAL,
                                 JsonTypeInfo.As.PROPERTY);
 
@@ -66,63 +69,50 @@ public class RedisConfig {
                                 .serializeValuesWith(
                                                 RedisSerializationContext.SerializationPair.fromSerializer(serializer))
                                 .disableCachingNullValues()
-                                .entryTtl(Duration.ofMinutes(30)); // TTL por defecto
+                                .entryTtl(Duration.ofMinutes(30));
 
                 // Configuraciones específicas por tipo de caché
                 Map<String, RedisCacheConfiguration> cacheConfigurations = new HashMap<>();
 
-                // Products and Recipes Pages (v4)
+                // ... (rest of configuration unchanged)
                 cacheConfigurations.put("products_page_v4", defaultConfig.entryTtl(Duration.ofMinutes(10)));
                 cacheConfigurations.put("recipes_page_v4", defaultConfig.entryTtl(Duration.ofMinutes(10)));
-
-                // Products: 1 hora (se modifican con frecuencia)
                 cacheConfigurations.put("product_v4", defaultConfig.entryTtl(Duration.ofHours(1)));
-
-                // Recipes: 2 horas (más estables)
                 cacheConfigurations.put("recipe_v4", defaultConfig.entryTtl(Duration.ofHours(2)));
-
-                // Users: 30 minutos (datos de autenticación/autorización)
                 cacheConfigurations.put("users", defaultConfig.entryTtl(Duration.ofMinutes(30)));
                 cacheConfigurations.put("user", defaultConfig.entryTtl(Duration.ofMinutes(30)));
                 cacheConfigurations.put("userByEmail", defaultConfig.entryTtl(Duration.ofMinutes(30)));
                 cacheConfigurations.put("userDetails", defaultConfig.entryTtl(Duration.ofMinutes(15)));
-
-                // Orders: 15 minutos (datos transaccionales)
                 cacheConfigurations.put("orders", defaultConfig.entryTtl(Duration.ofMinutes(15)));
                 cacheConfigurations.put("order", defaultConfig.entryTtl(Duration.ofMinutes(15)));
-
-                // Allergens: 24 horas (datos maestros que raramente cambian)
                 cacheConfigurations.put("allergens", defaultConfig.entryTtl(Duration.ofHours(24)));
                 cacheConfigurations.put("allergen", defaultConfig.entryTtl(Duration.ofHours(24)));
-
-                // Recipe components y allergens: 2 horas
                 cacheConfigurations.put("recipeComponents", defaultConfig.entryTtl(Duration.ofHours(2)));
                 cacheConfigurations.put("recipeAllergens", defaultConfig.entryTtl(Duration.ofHours(2)));
 
                 return RedisCacheManager.builder(connectionFactory)
                                 .cacheDefaults(defaultConfig)
                                 .withInitialCacheConfigurations(cacheConfigurations)
-                                .transactionAware() // Soporte para transacciones
+                                .transactionAware()
                                 .build();
         }
 
-        /**
-         * RedisTemplate para operaciones manuales con Redis si son necesarias.
-         */
         @Bean
-        public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory connectionFactory) {
+        public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory connectionFactory,
+                        ObjectMapper objectMapper) {
                 RedisTemplate<String, Object> template = new RedisTemplate<>();
                 template.setConnectionFactory(connectionFactory);
 
-                // Serialización de claves como String
                 template.setKeySerializer(new StringRedisSerializer());
                 template.setHashKeySerializer(new StringRedisSerializer());
 
-                // Serialización de valores como JSON
-                ObjectMapper objectMapper = new ObjectMapper();
-                objectMapper.registerModule(new JavaTimeModule());
+                // Configuración de validación de tipos polimórficos
+                PolymorphicTypeValidator ptv = BasicPolymorphicTypeValidator.builder()
+                                .allowIfBaseType(Object.class)
+                                .build();
+
                 objectMapper.activateDefaultTyping(
-                                LaissezFaireSubTypeValidator.instance,
+                                ptv,
                                 ObjectMapper.DefaultTyping.NON_FINAL,
                                 JsonTypeInfo.As.PROPERTY);
 
