@@ -4,7 +4,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
-import org.springframework.security.test.context.support.WithMockUser;
 
 import com.economato.inventory.dto.response.RecipeAuditResponseDTO;
 
@@ -38,12 +37,12 @@ class RecipeAuditControllerIntegrationTest extends BaseControllerMockTest {
     }
 
     @Test
-    @WithMockUser(username = "admin", roles = { "ADMIN" })
+    
     void getAllRecipeAudits_ShouldReturnList() throws Exception {
 
         when(recipeAuditService.findAll(any(Pageable.class))).thenReturn(testRecipeAudits);
 
-        mockMvc.perform(get("/api/recipe-audits")
+        mockMvc.perform(get("/api/recipe-audits").with(org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user("admin").roles("ADMIN"))
                 .param("page", "0")
                 .param("size", "10")
                 .contentType(MediaType.APPLICATION_JSON))
@@ -55,12 +54,12 @@ class RecipeAuditControllerIntegrationTest extends BaseControllerMockTest {
     }
 
     @Test
-    @WithMockUser(username = "admin", roles = { "ADMIN" })
+    
     void getRecipeAuditById_WhenExists_ShouldReturnAudit() throws Exception {
 
         when(recipeAuditService.findById(1)).thenReturn(Optional.of(testRecipeAudit));
 
-        mockMvc.perform(get("/api/recipe-audits/1")
+        mockMvc.perform(get("/api/recipe-audits/1").with(org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user("admin").roles("ADMIN"))
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id_recipe").value(1))
@@ -69,23 +68,23 @@ class RecipeAuditControllerIntegrationTest extends BaseControllerMockTest {
     }
 
     @Test
-    @WithMockUser(username = "admin", roles = { "ADMIN" })
+    
     void getRecipeAuditById_WhenNotExists_ShouldReturn404() throws Exception {
 
         when(recipeAuditService.findById(anyInt())).thenReturn(Optional.empty());
 
-        mockMvc.perform(get("/api/recipe-audits/999")
+        mockMvc.perform(get("/api/recipe-audits/999").with(org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user("admin").roles("ADMIN"))
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(username = "admin", roles = { "ADMIN" })
+    
     void getAuditsByRecipeId_ShouldReturnList() throws Exception {
 
         when(recipeAuditService.findByRecipeId(1)).thenReturn(testRecipeAudits);
 
-        mockMvc.perform(get("/api/recipe-audits/by-recipe/1")
+        mockMvc.perform(get("/api/recipe-audits/by-recipe/1").with(org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user("admin").roles("ADMIN"))
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$").isArray())
@@ -93,12 +92,12 @@ class RecipeAuditControllerIntegrationTest extends BaseControllerMockTest {
     }
 
     @Test
-    @WithMockUser(username = "admin", roles = { "ADMIN" })
+    
     void getAuditsByUserId_ShouldReturnList() throws Exception {
 
         when(recipeAuditService.findByUserId(1)).thenReturn(testRecipeAudits);
 
-        mockMvc.perform(get("/api/recipe-audits/by-user/1")
+        mockMvc.perform(get("/api/recipe-audits/by-user/1").with(org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user("admin").roles("ADMIN"))
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$").isArray())
@@ -106,13 +105,13 @@ class RecipeAuditControllerIntegrationTest extends BaseControllerMockTest {
     }
 
     @Test
-    @WithMockUser(username = "admin", roles = { "ADMIN" })
+    
     void getAuditsByDateRange_ShouldReturnList() throws Exception {
 
         when(recipeAuditService.findByMovementDateBetween(any(LocalDateTime.class), any(LocalDateTime.class)))
                 .thenReturn(testRecipeAudits);
 
-        mockMvc.perform(get("/api/recipe-audits/by-date-range")
+        mockMvc.perform(get("/api/recipe-audits/by-date-range").with(org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user("admin").roles("ADMIN"))
                 .param("start", "2026-01-01T00:00:00")
                 .param("end", "2026-02-01T23:59:59")
                 .contentType(MediaType.APPLICATION_JSON))
